# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OMPAY Orange Money API",
        "description": "API pour l'application Orange Money - OMPAY",
        "contact": {
            "email": "support@orangemoney.sn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        },
        {
            "url": "https://api.ompaysn.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/portefeuille/solde": {
            "get": {
                "tags": [
                    "Portefeuille"
                ],
                "summary": "Consulter le solde du portefeuille",
                "description": "Récupère le solde actuel du portefeuille de l'utilisateur",
                "operationId": "consulterSolde",
                "responses": {
                    "200": {
                        "description": "Solde récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "solde": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 91500
                                                },
                                                "devise": {
                                                    "type": "string",
                                                    "example": "XOF"
                                                },
                                                "dateDerniereMiseAJour": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-10T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/portefeuille/transactions": {
            "get": {
                "tags": [
                    "Portefeuille"
                ],
                "summary": "Historique des transactions",
                "description": "Récupère l'historique des transactions de l'utilisateur avec pagination et filtres",
                "operationId": "historiqueTransactions",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de la page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "limite",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type de transaction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "transfert",
                                "paiement",
                                "tous"
                            ]
                        }
                    },
                    {
                        "name": "dateDebut",
                        "in": "query",
                        "description": "Date de début (format Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "dateFin",
                        "in": "query",
                        "description": "Date de fin (format Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Statut de la transaction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en_attente",
                                "termine",
                                "echoue",
                                "annule"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historique récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "transactions": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "example": "uuid"
                                                            },
                                                            "type": {
                                                                "type": "string",
                                                                "example": "transfert"
                                                            },
                                                            "montant": {
                                                                "type": "number",
                                                                "example": 5000
                                                            },
                                                            "devise": {
                                                                "type": "string",
                                                                "example": "XOF"
                                                            },
                                                            "statut": {
                                                                "type": "string",
                                                                "example": "termine"
                                                            },
                                                            "dateCreation": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-11-10T12:00:00Z"
                                                            },
                                                            "reference": {
                                                                "type": "string",
                                                                "example": "TRF123456789"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "limite": {
                                                            "type": "integer",
                                                            "example": 20
                                                        },
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 150
                                                        },
                                                        "pages": {
                                                            "type": "integer",
                                                            "example": 8
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/portefeuille/transactions/{idTransaction}": {
            "get": {
                "tags": [
                    "Portefeuille"
                ],
                "summary": "Détails d'une transaction",
                "description": "Récupère les détails complets d'une transaction spécifique",
                "operationId": "detailsTransaction",
                "parameters": [
                    {
                        "name": "idTransaction",
                        "in": "path",
                        "description": "ID de la transaction",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la transaction récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "uuid"
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "example": "transfert"
                                                        },
                                                        "montant": {
                                                            "type": "number",
                                                            "example": 5000
                                                        },
                                                        "frais": {
                                                            "type": "number",
                                                            "example": 100
                                                        },
                                                        "devise": {
                                                            "type": "string",
                                                            "example": "XOF"
                                                        },
                                                        "statut": {
                                                            "type": "string",
                                                            "example": "termine"
                                                        },
                                                        "reference": {
                                                            "type": "string",
                                                            "example": "TRF123456789"
                                                        },
                                                        "dateCreation": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2025-11-10T12:00:00Z"
                                                        },
                                                        "dateFinalisation": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2025-11-10T12:05:00Z"
                                                        },
                                                        "destinataire": {
                                                            "properties": {
                                                                "nom": {
                                                                    "type": "string",
                                                                    "example": "Admin Orange Money"
                                                                },
                                                                "numeroTelephone": {
                                                                    "type": "string",
                                                                    "example": "+221771234567"
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "marchand": {
                                                            "properties": {
                                                                "nom": {
                                                                    "type": "string",
                                                                    "example": "Boutique Express"
                                                                },
                                                                "numeroMarchand": {
                                                                    "type": "string",
                                                                    "example": "MERCHANT123"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "TRANSACTION_NOT_FOUND"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Transaction non trouvée"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}